+++ modules/java.sh
@@ -3,15 +3,14 @@
 
 enterDirOrExit "${CURRENT_DIR}"
 
-if grep "JAVA_HOME" ~/.zshrc; then
-   logAlreadyInstalled "Java"
-   logProgramVersion "Java" "$(java version)"
-else
-    logInfo "Download latest Java7 from: http://www.oracle.com/technetwork/java/javase/downloads/java-archive-downloads-javase7-521261.html"
-    logInfo "Download latest Java8 from: http://www.oracle.com/technetwork/java/javase/downloads/java-archive-javase8-2177648.html"
-    logInfo "Remember to update JAVA_HOME environment variable after installing Java"
+if ! command -v java >/dev/null; then
+    preInstallationLog "Java"
 
-    log "Setting up Java environment variables..."
-    addToShell "export JAVA_HOME=/Library/Java/JavaVirtualMachines/jdk1.7.0_79.jdk/Contents/Home"
-    addToShell "export PATH=\$JAVA_HOME/bin:\$PATH"
+    sdk install java
+
+    postInstallationLog "Java"
+else
+    logAlreadyInstalled "Java"
 fi
+
+logProgramVersion "Java" "$(java -version)"